refactor: move reply_desc into its own file
it's grown up enough now to deserve that at least
